move test_molecule to helper_functions
