+++ vermouth/tests/helper_functions.py
@@ -17,6 +17,8 @@ Contains helper functions for tests.
 """
 import operator
 import os
+import pytest
+import numpy as np
 import networkx as nx
 import networkx.algorithms.isomorphism as iso
 import vermouth
@@ -148,3 +150,56 @@ def create_sys_all_attrs(molecule, moltype, secstruc, defaults, attrs):
     system = System()
     system.molecules.append(molecule)
     return system
+
+@pytest.fixture
+def test_molecule(scope='function'):
+    """
+    Molecule with the following connectivity and atom-naming:
+
+    SC2:   2           8
+           |           |
+    SC1:   1   4       7
+           |   |       |
+    BB:    0 - 3 - 5 - 6
+           -------------
+    resid: 1   2   3   4  column wise
+    """
+
+    force_field = vermouth.forcefield.ForceField("test")
+    molecule = vermouth.molecule.Molecule(force_field=force_field)
+    molecule.meta['test'] = True
+    # The node keys should not be in a sorted order as it would mask any issue
+    # due to the keys being accidentally sorted.
+    molecule.add_node(2, atomname='SC2',
+                      position=np.array([0., 1.0, 0.0]), resid=1)
+    molecule.add_node(0, atomname='BB',
+                      position=np.array([0., 0., 0.]), resid=1)
+    molecule.add_node(1, atomname='SC1',
+                      position=np.array([0., 0.5, 0.0]), resid=1)
+
+    molecule.add_node(3, atomname='BB', position=np.array(
+        [0.5, 0.0, 0.0]), resid=2)
+    molecule.add_node(4, atomname='SC1', position=np.array(
+        [0.5, 0.5, 0.0]), resid=2)
+
+    molecule.add_node(5, atomname='BB', position=np.array(
+        [1.0, 0.0, 0.0]), resid=3)
+
+    molecule.add_node(6, atomname='BB', position=np.array(
+        [1.5, 0.0, 0.0]), resid=4)
+    molecule.add_node(7, atomname='SC1', position=np.array(
+        [1.5, 0.5, 0.0]), resid=4)
+    molecule.add_node(8, atomname='SC2', position=np.array(
+        [1.5, 1.0, 0.0]), resid=4)
+
+    molecule.add_edge(0, 1)
+    molecule.add_edge(0, 2)
+    molecule.add_edge(0, 3)
+    molecule.add_edge(3, 4)
+    molecule.add_edge(3, 5)
+    molecule.add_edge(5, 6)
+    molecule.add_edge(6, 7)
+    molecule.add_edge(7, 8)
+
+    return molecule
+
